feat: PhysarumSolverの抽象化
